third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-3/styles-cancel-editing.html

<html>
<head>
<script src="../../../inspector/inspector-test.js"></script>
<script src="../../../inspector/elements-test.js"></script>
<script>

function test() {
  ElementsTestRunner.selectNodeAndWaitForStyles('inspected', step1);

  var treeElement;
  var section;

  function step1() {
    ElementsTestRunner.dumpSelectedElementStyles(true);
    treeElement = ElementsTestRunner.getElementStylePropertyTreeItem('color');

    treeElement.startEditing();
    treeElement.nameElement.textContent = 'color';
    treeElement.nameElement.dispatchEvent(TestRunner.createKeyEvent('Enter'));

    // Update incrementally, do not commit.
    treeElement.valueElement.textContent = 'green';
    treeElement.kickFreeFlowStyleEditForTest();

    // Cancel editing.
    treeElement.valueElement.firstChild.select();
    ElementsTestRunner.waitForStyleApplied(onStyleApplied);

    function onStyleApplied() {
      treeElement.valueElement.dispatchEvent(TestRunner.createKeyEvent('Escape'));
      ElementsTestRunner.waitForStyleApplied(onStyleReverted);
    }

    function onStyleReverted() {
      ElementsTestRunner.selectNodeWithId('other', step2);
    }
  }

  function step2() {
    ElementsTestRunner.selectNodeAndWaitForStyles('inspected', step3);
  }

  function step3() {
    TestRunner.addResult('After append:');
    ElementsTestRunner.dumpSelectedElementStyles(true);
    TestRunner.completeTest();
  }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that editing is canceled properly after incremental editing.
</p>

<div id="inspected" style="color: red">Text</div>
<div id="other"></div>

</body>
</html>
