src/third_party/blink/web_tests/http/tests/devtools/elements/styles-3/styles-cancel-editing-expected.txt

Tests that editing is canceled properly after incremental editing.

[expanded] 
element.style { ()
    color: red;

[expanded] 
div { (user agent stylesheet)
    display: block;

======== Inherited from html ========
[expanded] 
html { (user agent stylesheet)
/-- overloaded --/     color: -internal-root-color;

After append:
[expanded] 
element.style { ()
    color: red;

[expanded] 
div { (user agent stylesheet)
    display: block;

======== Inherited from html ========
[expanded] 
html { (user agent stylesheet)
/-- overloaded --/     color: -internal-root-color;
